Create index.html
This is going to be a html editor with a DOM-tree wiever on it's left side, and an easy copyable tag-list on it's right side. The focus-handling of the right-side is done: it's enough to place the cursor once over the right box, and from that the marks of the list is following the movement of the cursor.